Updated the design and removed redundant code in the index.html, style.css,script.js, thank-you-page.html and privacy-policy-page.html files.
Removed the meta name title element in the index.html, thank-you-page.html and privacy-policy-page.html files, the screen reader repeats it and it is not a necessary element to have.
Changed the footer absolute link content, to make it more simple for users to understand in the index.html file.
Changed the portfolio section elements class name, to make it more readable in the index.html file.
Removed the redundant aria-label in the laptop image SVG container as it has aria-hidden true in the index.html file.
Changed the SVG image and added new paths to make it look more realistic and detailed in the index.html file.
Changed the style for the landing page title to help it stand out more in the style.css file.
Changed the position of the landing page title and laptop image SVG container in the style.css file.
Changed the padding in the portfolio section container and portfolio websites container to make it consistent with the other sections in the style.css file.
Changed the style and added the dark mode selectors for the laptop image SVG image in the style.css file.
Changed the order of the HTML elements, CSS selectors and JavaScript dark mode function for the laptop image SVG to make it more readable, in the index.html, style.css and script.js files.
Changed the contact form label event listener to focus instead of input to help users who use the contrast settings. It enables the label animation on focus and allows the placeholder to be visible without the label in the way visually in the script.js file.
Commented out some code ready for full tests before it is removed.