Updated DRY, removed redundant code and compressed a font in the style.css file
Changed the repeating CSS declaration  blocks in the style.css file.
Removed any code that was unnecessary  in the style.css file.
Added a compressed font in the style.css file.
Removed two fonts that are redundant in the style.css file.
Cleaned up the code and made it consistent in the script.js file.