Updated readability and DRY
Removed any unnecessary and redundant code in the index.html and style.css files.
Removed an unused font file.
Moved hidden hidden screen reader comments to make it more readable in the index.html file.
Added a span for the highlighted text in the portfolio and contact section, this is to help with user select, DRY and screen readers in the index.html file.
Prevented repeating code which helped the DRY in the style.css file.
Moved the CSS order to make it more readable and organized in the style.css file.
Cleaned up the selectors in the style.css file.