Updated the order and separation of the CSS files.
Separated the CSS files into responsive, dark mode and modal to make the CSS more readable and organised.
Changed the order of the CSS in the files and added a documentational comment on each file, in order for other developers to understand the order of the CSS.
Renamed the privacy policy page CSS file to make it consistent.
Added the stylesheet links to the HTML files.