Updated the portfolio contact form and portfolio privacy policy page.
Added aria attributes to the contact form elements, which made it so the error messages are read out by the screen reader on blur and the error message appears in the index.html file.
Added a max length attribute to the phone input box so that users are not able to type in more numbers than what is needed in the index.html file.
Changed the image rendering value from smooth because it is only compatible with Firefox and no other modern browsers in the style.css file.
Changed the script for the portfolio contact form by adding a contact form screen reader message announcement on blur. This is to let the user know that they need to enter a message before sending the form in the script.js file.
Added a function to announce the error messages when they appear on blur, on the portfolio contact form in the script.js file.
Updated the privacy policy page by adding the same top navigation bar and footer navigation container elements as the portfolio and adding the same styles and functionality (including dark mode) in the privacy-policy-page.html, privacy-policy-page-style.css and privacy-policy-page-script.js files.