Updated the thank-you-page HTML and CSS styles for the portfolio, privacy-policy-page and thank-you-page
A new folder and file appeared in VSCode, it is to be able to view the website in live server.
Added a div with the class of form-background to the contact form, to be able to add a background and style it in the index.html file.
Added a box-shadow to the portfolio profile container, portfolio container, contact form container and footer navigation container in the style.css file.
Added styling to the form-background in the style.css file.
Removed the left and right borders from the portfolio contact form input boxes in the style.css file.
Added a focus pseudo class to the privacy-policy-page-content-container a and grouped the hover and focus pseudo selectors in the privacy-policy-page.css file.
Removed redundant code in the privacy-policy-page.js file.
Changed the functionality of the dark mode cookie in the privacy-policy-page.js file.
Added a comment to fix the functionality for the top navigation bar not opening and closing when pressing space or enter in the privacy-policy-page.js file.
Added the top navigation bar HTML code to make it the same as the index.html in the thank-you-page.html file.
Added the top of the page accessibility label HTML code to make it the same as the index.html in the thank-you-page.html file.
Changed the section title from capital letters to capitalised in the thank-you-page.html file.
Added a tab index of minus one, and added icons to the buttons in the thank-you-page.html file.
Added the footer HTML code to make it the same as the index.html in the thank-you-page.html file.
Changed the styles for the thank-you-page.css file.
Added the same functionality as the script.js and privacy-policy-page.js files for the thank-you-page.js file.