Fixed styling bug and updated accessibility code in the privacy-policy-page and thank-you-page
Changed the contents navigation anchor link, element content for accessibility in the privacy-policy-page.html file.
Added aria-hidden to the SVG images in the thank-you-page.html file to improve accessibility.
Fixed the parallax image, filter blur showing on the top of the section title header slightly, in the thank-you-page.css and privacy-policy-page.css files.