Updated: footer bug fixes, font styles, dark mode styles, transition styles, background height issues, removed unnecessary code and comments, positioning of elements, hamburger menu accessibility issues, button focus accessibility issue, accessibility anchoring issue.
Removed any unnecessary code and comments.

Changed the message sent paragraph content in the thank-you-page.html file.

Fixed the SVG path animation issue in the thank-you-page.html file.

Fixed the focus issue for accessibility in the thank-you-page.html file.

Changed the font styles for the element content that had the letter i and the font family, roboto bold, as the font was not readable and negatively impacted the user experience and interface in the style.css, thank-you-page-style.css and privacy-policy-page.css files.

Fixed the height issue of the form background in the style.css file.

Changed the hyperlinks transition background and time in the style.css file.

Fixed the footer positioning on the style.css, thank-you-page-style.css and privacy-policy-page-style.css files.

Fixed the anchor accessibility issue for the contents navigation, in the privacy-policy-page.html and privacy-policy-page-script.js files.

Changed the welcome page title script adding a .forEach() method to make sure the span was also included when dark mode was toggled in the script.js file.

Added the dark mode colour and functionality for the contact forms dark mode background in the style.css and script.js files.

Fixed the scripts in order for the hamburger menu to be collapsed and expanded using not only click, but enter and space as well in the script.js, thank-you-page-script.js and privacy-policy-page.js files.