Updated the readme, portfolio, thank you page and privacy policy files
Added details of accessibility features in the README.md file.

Changed the file names of the parallax image, headshot image, landing page image and landing page dark mode image in the index.html and style.css files.

Enhanced the headshot image on the index.html file.

Made the quotation container more maintainable in the profile section in the index.html, style.css and dark-mode-style.css files.

Changed the path to the thank you page for testing in the index.html file.

Changed the variable values to make the elements assigned, more specific in the script.js, thank-you-page.js and privacy-policy-page.js files.

Changed the conditions in the dark mode and hCaptcha modals if statements to make it more maintainable in the script.js, thank-you-page-script.js and privacy-policy-page.js files.

Added the portfolio quotation containers dark mode functionality in the script.js file.

Cleaned up indentation and comments in the script.js file.

Changed the contact form submission variable name to make it more meaningful in the script.js file.

Added an animation to the top navigation bar when the user clicks on the hamburger menu in the style.css, script.js, thank-you-page-script.js and privacy-policy-page.js files.

Changed the style of the website slider buttons to make them stand out more and become less plain in the style.css and responsive-style.css files.

Changed some of the CSS selectors in the style.css file.